refactor : separate page data into standalone files; refactor image referencing methods and vh style settings
