+++ src/components/common/LoadingSpinner.vue
@@ -1,10 +1,6 @@
-<style lang="scss" scoped>
-.vh--80 {
-  height: 80vh;
-}
-</style>
+<style lang="scss" scoped></style>
 <template>
-  <div class="vh--80 d-flex justify-content-center align-items-center">
+  <div class="vh-80 d-flex justify-content-center align-items-center">
     <div class="spinner-border text-warning" role="status">
       <span class="visually-hidden">Loading...</span>
     </div>
