+++ src/components/front/Index/EvaluationBoard.vue
@@ -49,7 +49,7 @@
 <template>
   <div
     class="evaluation-board__wrap"
-    :style="{ backgroundImage: `url(${evaluationBoardInfo.bgImg})` }"
+    :style="{ backgroundImage: `url(${getImageUrl(evaluationBoardInfo.bgImg)})` }"
   >
     <div class="evaluation-board__container">
       <div class="container">
@@ -122,15 +122,15 @@
 <script setup>
 import { computed } from 'vue'
 import { RouterLink } from 'vue-router'
+import { useMediaQuery } from '@vueuse/core'
+import { evaluationData } from '@/data/front/evaluationData.js'
+import { getImageUrl } from '@/utils/imgUrl.js'
 import IconTitle from '@/components/front/base/IconTitle.vue'
 
-import { getImageUrl } from '@/utils/imgUrl.js'
-import bgImg from '@/assets/bg-img/shaba.jpeg'
-
-import { useMediaQuery } from '@vueuse/core'
 const isMediaLgUp = useMediaQuery('(min-width: 992px)')
 
 const evaluationBoardInfo = {
+  bgImg: 'assets/bg-img/shaba.jpeg',
   title: {
     isClock: false,
     title: '郊友真情推薦',
@@ -141,7 +141,6 @@ const evaluationBoardInfo = {
     }
   },
   subtitle: '不知如何走出第一步？',
-  bgImg,
   text: {
     main: '郊友帶你博覽群山',
     vice: '尋訪與自然的相會'
@@ -153,48 +152,4 @@ const titleTextColor = computed(() =>
     ? evaluationBoardInfo.title.textColor.mediaLgUp
     : evaluationBoardInfo.title.textColor.normal
 )
-
-const evaluationData = [
-  {
-    user: '賞鳥粉絲',
-    avatar: 'assets/avatar/bird.jpeg',
-    evaluate: '★★★★☆',
-    mes: '沿線植物資源豐富、鳥況極佳，以中低海拔鳥類如五色鳥、藪鳥、山紅頭、冠羽畫眉等為主。',
-    trail_data: {
-      TR_CNAME: '福山步道',
-      TRAILID: '178'
-    }
-  },
-  {
-    user: '鐵道迷',
-    avatar: 'assets/avatar/train.jpeg',
-    evaluate: '★★★★★',
-    mes: '全台獨一無二與鐵道相依的登山步道，可從不同的方向三次俯瞰樟腦寮村莊，親自感受獨立山螺旋狀三迴旋之美。',
-    trail_data: {
-      TR_CNAME: '獨立山步道',
-      TRAILID: '085'
-    }
-  },
-  {
-    user: '百岳發燒友',
-    avatar: 'assets/avatar/hiker.jpeg',
-    evaluate: '★★★☆☆',
-    mes: '中央山脈主脊南段，為南橫三星中最易攀登者，登頂後為眺望向陽山南面大崩壁最佳觀賞處。',
-    trail_data: {
-      TR_CNAME: '關山嶺山步道',
-      TRAILID: '148'
-    }
-  },
-  {
-    user: '賞蝶愛好',
-    avatar: 'assets/avatar/butterfly.jpeg',
-    evaluate: '★★★☆☆',
-    mes: '步道為低海拔天然闊葉林，沿途蜜源植物更吸引許多蝶類、蜜蜂，生物資源豐富。',
-    trail_data: {
-      TR_CNAME: '鎮南宮步道',
-      TRAILID: '174'
-    }
-  }
-]
 </script>
-@/utils/imgUrl
